requirements.txt
google-cloud-storage
sqlalchemy
pandas
dash
plotly
python-dotenv
psycopg2
schedule